Fix AIX build.
